+++ examples/server/src/modules/ws-todo/ws.d.ts
@@ -1,3 +1,0 @@
-interface ExtWebSocket extends WebSocket {
-  isAlive: boolean;
-}